Add initial project creation files
In order to use git source control for our new example blog project
that's going to make millions, this commit adds all of the rails
generated files as well as a bit of basic functionality. It includes
crud operations for posts, tags, and comments. It also includes showing
authors, but no other crud operations.